+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/ControllerScanner.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.mvc;
 
 import java.lang.reflect.InvocationTargetException;
 import java.util.HashMap;
